Update v1.3
The system of saving
The problem of limits on cloud saving has been solved. Now the data is stored both locally and on the cloud. Everything happens automatically, no action is required from you. There are new saving options in InfoYG.

ViewingAds
Added cursor control. It is recommended to check the ViewingAds script settings in your project after the update. Useful: you can use several ViewingAds components at once in order to control different parameters when opening/closing ads.

Time type for leaderboards
Finalized! The documentation has been updated.

Image Load YG
The component has got the Sprite Image option. Drag the object with the Image component to the Sprite Image so that the image loads as a sprite.

Added a static event YandexGame.onResetProgress — this is the setting of all parameters and saves of the game by default. Called at the first start of the game.

Added a static event GraphicSettingsYG.onQualityChange — called when changing graphics by means of the GraphicSettingsYG component.

Replaced the standard logo when loading the game with the plugin logo. Still, don't use it. Use your logo.

The new syntax "new()" is returned to the old "new Type()". There will be no more errors on older software versions because of this.