Update 1.3.1
Updating documentation
Including the new section "Frequently Asked questions"

Fast implementation of Yandex. Metrica
Create a counter on the Yandex.Metrica website (https://metrika .yandex.ru/list?). Copy the counter code. Find the "Metric" field in the PluginYG template. Insert the counter code into this field.

Fix the game freezes on a mobile device when opening a dialog box
Now there is no hanging, except for the evaluation window of the game! (Hang-up window games are tested in Yandex browser and Google Chrome. There is no hang in Chrome, the game crashes in Yandex!).

Fix (https://t.me/yandexgame_plugin/13109 ) problems with loading cloud saves when using json in saved data.

Fix (https://t.me/yandexgame_plugin/12681 ) a bug with adding an icon to the desktop. Now, when you close the window offering to create an icon, the button for calling such a window in the game is hidden, since Yandex does not allow you to open such a window twice.