Update v1.3.2
To switch from previous versions of the plugin to v1.3.2, you may need to reconfigure the ViewingAdsYG component and the option to display ads when loading scenes!

- Automatic build archiving
After the successful creation of the game build, the folder with the contents of the build is packed into a zip archive. When reassembling the game, the archive will not be overwritten, but a new package will be created with the assigned number in the file name.

- Updated the Viewing Ads YG script
The documentation has been updated. When switching from previous versions of the plugin, you need to make sure that the settings of the Viewing Ads YG component are satisfactory, because some options have been changed.

- Adjusted the function of displaying ads when loading scenes
In the settings of the InfoYG plugin, the option to display ads when loading scenes has changed. Now this option is called Ad When Loading Scene. By default = true — this means that the ad display will be called when loading any scene in the game. The value is false — ads will not be displayed when loading scenes.

- Disabling the display of full-screen ads when starting the game
By default, the first advertisement is shown even before the game is loaded, this can be disabled:
1. Go to the Project Settings Project Settings → Player → Resolution and Presentation → WebGL Template
2. Select the PluginYG template
3. Find the field Off ads before loading game
4. Fill in this field with any word, for example, write true there
5. Now full-screen ads will not be displayed immediately when the game starts.

- Adjusted the display of ads when launching the game for different devices
1. On Desktop, ads are displayed immediately when the page is opened.
2. On mobile devices, ads are shown only after the game is downloaded. The instant display of ads after opening the page with the game can be annoying, especially if the advertising slows down, and because of this, it will also be difficult to close it, which may happen just when the game is loading. For this reason, ads will be displayed on mobile devices after the game is loaded.

- The button to open the evaluation window of the game
The button for calling the game evaluation window is now disabled if the game is running on a mobile device, because at the moment, calling such a window for some reason can disable the game in some of the browsers.
You can turn the button back on, the Show On Mobile Device option in the Review YG component is responsible for this.

- Fix for ESC
When working with ECS frameworks (DOTS, LeoECS, Morpeh, etc...), domain reboot shutdown is often used. For this option, a reboot of the plugin's static fields has been implemented, which should ensure correct operation with ECS.

- Fix for incognito mode
In previous versions of the plugin, in incognito mode in certain browsers, loading local saves occurred with an error due to the restriction of browser access to data, and then due to an error, the plugin did not work correctly. Starting from the plugin version 1.3.2, this error is successfully handled.

- Fix for the Language YG component
In previous versions of the plugin, if there was no YandexGame object on the scene, the LanguageYG component could not detect the Scriptable Object “InfoYG". Starting from the plugin version 1.3.2, if there is no YandexGame object on the stage, then InfoYG is taken from the YandexGame → WorkingData folder.