PluginYG v1.3.3
- Ability to disable automatic build archiving
In the Info plugin settings, the Auto Build Archiving option is responsible for the activity of the build autoarchivation function.

- Fix a bug in which the event of closing a full-screen ad could not be executed during simulation in Unity
If a new scene was loaded at the time of the simulation of the display of a full-screen ad, then the ad closing event could not be executed, because of this, the ad was not shown further. This bug was present only in Unity Editor, but not in the finished build.