v1.3.5
- SDK initialization logic changed
When testing a draft in the “turtle” mode (limiting query timeouts), player data will not be loaded (it will be like an unauthorized user), except for saves, they will be taken from local saves.

- Implemented “Start of games”
Signaling to Yandex that the game has loaded.

- Setting the Pixel Ratio
In the PluginYG template, the Pixel ratio mobile field is responsible for the value of the Device Pixel Ratio in the index file.
Fill in the field with the value you selected, for example: 1.3
Leave this field empty if you don't want to change anything.

- Cut out the sitelock
In connection with the new paragraph 1.18 in the rules of requirements for the game the site lock was cut because it blocked the game page located on a different domain.

- Adjusted the display of ads when launching the game for different devices
In the v1.3.2 update, there was an item of the same name. Now it's the same as it was before the v1 update.3.2 due to a change in the SDK initialization logic.

- Fixed the game hang
On mobile devices, in some browsers, when opening some windows and clicking on the Ya game curtain, the game hung. This shouldn't be happening now.

- Saving an array of arrays
Now you can choose between Json Utility and JsonConvert.
The disadvantage of the JsonConvert class is that when using it, the JsonNet library is added to the project build and the final weight of the game increases by ~ 2mb.
The advantage of the JsonConvert class is saving an array in an array. Use JsonUtility if you don't need to save an array in an array, nothing needs to be done for this, JsonUtility is used by default.
To use JsonConvert: uncomment the first prompt “//#define JSON_NET_ENABLED” in the YandexGame script.

- Implemented translation of the words “unauthorized” and “anonymous”
Translation into all languages is embedded in the scripts “LeaderboardYG” and “GetPlayerYG".
In your own scripts, when receiving data, for example, the user name, if the name is “unauthorized” or “anonymous”, then it will not be translated. That is, in your scripts, the translation will need to be done independently. To do this, you can use the methods:
YandexGame.Instance.infoYG.UnauthorizedTextTranslate();
YandexGame.Instance.infoYG.IsHiddenTextTranslate();
The word anonymous has been replaced by Is Hidden — meaning “hidden".
The methods return the translated word. The language is determined depending on the plugin settings. The methods also have a language overload — You can write the language into which you want to translate the word into an overload.

- The CloseFullAdEvent event has been finalized
Now it is called only after the successful display and closing of the ad.

- Added ErrorFullAdEvent event
ErrorFullAdEvent when a full-screen ad is displayed incorrectly.

- Improved the timer of the ad display interval
If when calling a full-screen ad, it was not shown due to an error or too frequent call, the timer of the interval for displaying the full-screen ad is reset.

- ReviewYG script
The game evaluation button will not be shown if the user is not logged in and the authDialog parameter = false. (If the user is not logged in and authDialog = true, the button will be shown and when the button is clicked, the authorization dialog box will open).

- OnURL_Yandex_DefineDomain method
If the domain is not defined for some reason, the link will not open.

- The LeaderboardYG component has got the "Is Hidden Player Photo" parameter
Specify a link to your sprite to display hidden users.

- The table update event in the LeaderboardYG component
"On Update Data" is called when receiving the leaderboard data.

- Getting user data from the leaderboard
The LeaderboardYG script has a PlayersData class - this class contains user data: name, rank, score, photo. The LeaderboardYG script contains the PlayerData[] field, which is an array whose elements are the PlayersData class described above. Accordingly, in the LeaderboardYG script, you can take an array of players and get their data. You can take data on the above-described event "On Update Data". A training script "GetLederboardData" was also created for an example of using data from a table, and in the demo scene this script is shown how it works. There is a "Leaderboard Debug Data" object in the scene. Pay attention to the leaderboard "Leaderboard Advanced & Time" in the scene (in the Canvas object) and its event "On Update Data", it is associated with the above script.

- Method for converting a record to a time type
In the LeaderboardYG script, you can find a method for converting a record to a time type: TimeTypeConvert(int);