PluginYG v1.5
- Localization
- Support for Text Mesh Pro
- New component Lang YG Additional Text.
- Autolocation now works on deactivated objects as well.

- Define system
In Info YG there are buttons for disabling and enabling:
-Newtonsoft libraries for using automatic localization.
- Libraries Newtonsoft for use in the save system. Newtonsoft json net allows you to save arrays in arrays, classes. This is advanced work with json. But in the build, the game will weigh 2mb more using Newtonsoft.
- Libraries TMPro (Text Mesh Pro).
Also, after importing the plugin, define YG_PLUGIN_YANDEX_GAME will be created in the scripting define symbols list. Library activity is also monitored using the scripting define symbols list.

- Advertising
- Added parameter:
`YandexGame.nowAdsShow`
`true' — Fullscreen or video ad is currently open.
`false' — Fullscreen or video ads are closed at the moment.

Earlier there were only parameters:
`YandexGame.nowFullAd`
`true' — Fullscreen ad is currently open.
`false' — The fullscreen ad is closed at the moment.

YandexGame.nowVideoAd`
`true' — Video advertising for a reward is open at the moment.
`false' — Video advertising for a reward is closed at the moment.

- If the time of viewing an advertisement for a reward is less than two seconds, then the reward will not be received. Instead, an error event will be called.
- Now, even with the Singleton option enabled, ads will be displayed when opening a new scene. Unless, of course, the Ad When Loading Scene parameter is enabled in Info YG.
- If an advertisement was opened when loading the game, and after loading it also remains open, then after initialization of the SDK, an advertising opening event is called.

- Additional control over the simulation
In Info YG you will find the simulation settings for:
- Player parameters. For example, the user's name, avatar, device.
- Creation of emulated leaderboards up to filling in information about players.
- The same for in-game purchases.
- Modification of game progress loading saves
- Added the Save On Quit Game option to Info YG. Executes the `SaveProgress` method when exiting the game (closing the page).
- The save file for testing in Unity Editor is now saved to the folder YandexGame/WorkingData/Editor and has the extension json. Now you can conveniently read and change the saves in the file.
- Fixed an issue where the compiler referred to the wrong error when saving or loading.

- Setting up the boot screen icon
In Info YG, the Logo Image Format parameter specifies the extension of the icon image, and the ability to disable it.

- New Activity On Authorization YG component
Allows you to perform an action depending on the user's authorization.
For example, it is easy to disable any button if the user is unauthorized.

- The update affected almost the entire plugin code
It is particularly possible to highlight:
- Leaderboards
- In-game purchases
For these items, see the updated documentation.
- Post process build system
- Game evaluation script
- Debugging tool
- Removing unnecessary
- Code restructuring
- Preparation for the modular system

- Links to documentation