PluginYG v1.6
Version 1.6
Initialization
The game will be downloaded only after the Yandex Games SDK is fully initialized.
Accessing plugin data is now optional. The initialization section becomes irrelevant.
The game starts on the local host.

Advertisement
New Ad Notification YG Script (documentation)
New Timer Script Before Ads YG (documentation)
In Info YG new parameter Load Ad With Delay Simulation:Delay in opening a full-screen ad. It can be useful for testing notifications that an ad will open soon, before it is shown (while waiting for an ad).
For testing in Unity Editor errors when displaying video ads, there is now a Test Error Of Rewarded Ads In Editor flag in Info YG. If you enable this option, a red screen will appear when you open an advertisement for a reward and no reward will be issued.

Improved the script Viewing Ads YG
Increased reliability. In particular, the script now deactivates the Event System during the ad display.
The Do Close Void On Awake parameter has been replaced with Awake Set Values.

Leaderboards
The record will be recorded in the table only if the user is logged in.
The ability not to write hidden users to the table. The Save Score Anonymous Players parameter in Info YG. true - anonymous users will be recorded. false - anonymous users will not be recorded.

Executing any method when closing the game
Executing any of your methods when closing or updating the game page. There are tooltips.

Updated documentation for metrics
More detailed information.

UTF-8 encoding
Not all scripts had the necessary UTF-8 encoding for the Russian language. Because of this, someone might not have displayed the Russian language. Now all scripts have UTF-8 encoding that supports Cyrillic.

Localization
Fixed some minor bugs with the Language YG and Lang YG Additional Text components.
Minor bug fixes

Version 1.5.2
of the Game Ready API
Now there is control over the Game Ready API. You can read about this in the new section.

Leaderboards
When simulating a leaderboard in Unity Editor, if when updating the leaderboard, the corresponding emulation leaderboard was not found by its name in Info YG, then an implicit error occurred earlier. Now the leaderboard is loading correctly, but it shows "no data".
In leaderboards, the sometimes inaccurate display of the decimal part of the score in the records inside the game has been fixed.

New documentation sections
Simulation of table display in Unity Editor
Simulation of displaying purchases in Unity Editor
Game Ready API

Version 1.5.1
Advertising
Disabling the first ad display in version 1.5 has broken down. In version 1.5.1, it works.
Fixed the possibility of game focus when showing the first ad at the moment when the game loaded.
Added the Rewarded After Closing parameter to Info YG. From my post in the Epic Shooter game, the second point of rejection was due to the fact that after the reward was issued, for some reason, the sound was played, but the advertisement was not closed yet. To exclude such a situation, the enabled Rewarded After Closing parameter will mean that the reward will be received only after viewing and closing the ad.
The YandexGame.timerShowAd field has become public. You can use it to track how much time is left until the next ad impression. This can be useful for a timer before showing ads.

Improved initialization processing
Added additional initialization checks. Now, even if initialization is not performed, data will be transmitted to the game as for an unauthorized user and Get Data Event will work. But the environment parameters will not be loaded.
In the script Viewing Ads YG the new parameter Do Close Void On Awake
Execute the Closing AD Values method in Viewing Ads YG in the Awake method (that is, at the start of the scene). This will allow you not to register events like audio pause = false or timeScale = 1 in your scripts in the Start methods.Pay attention to the new section of the documentation - “An easy way to configure Viewing Ads YG".

Additional checks for on/off custom banners
(For those who mistakenly compiled custom banners and got into the project, which could cause errors).