PluginYG 1.7
Updated the code related to changes in the Yandex Games SDK.

The openURL module has been added, which is completely similar to the module from the 2.0 version of the plugin.

To use the function associated with the Flags module in the openURL module, add the Flags_yg definition to the Scripting Define Symbols list (the list of definitions is located in Project Settings -> Player -> Other Settings).

In the Modules folder, you can delete the openURL and Flags modules if they are not used in the project. You can find the rest of the modules for version 1.x in the old documentation.